+++ fundamentals/src/org/fundamentals/homework/animals/Animal.java
@@ -1,4 +1,7 @@
-package org.fundamentals.homework;
+package org.fundamentals.homework.animals;
+
+import org.fundamentals.homework.animals.sounds.SoundMaker;
+import org.fundamentals.homework.animals.behavior.Behavior;
 
 public abstract class Animal implements Behavior, SoundMaker {
     public static final String DESCRIPTION = "%s says %s and %s";
