+++ fundamentals/src/org/fundamentals/homework/animals/Cat.java
@@ -1,4 +1,7 @@
-package org.fundamentals.homework;
+package org.fundamentals.homework.animals;
+
+import org.fundamentals.homework.animals.behavior.MiceCatcher;
+import org.fundamentals.homework.animals.sounds.MeowSound;
 
 public class Cat extends Animal implements MiceCatcher, MeowSound {
     public static final String NAME = "cat";
