+++ fundamentals/src/org/fundamentals/homework/animals/Dog.java
@@ -1,4 +1,7 @@
-package org.fundamentals.homework;
+package org.fundamentals.homework.animals;
+
+import org.fundamentals.homework.animals.behavior.HomeProtector;
+import org.fundamentals.homework.animals.sounds.BarkingSound;
 
 public class Dog extends Animal implements HomeProtector, BarkingSound {
     public static final String NAME = "dog";
