+++ fundamentals/src/org/fundamentals/homework/animals/Hedgehog.java
@@ -1,4 +1,7 @@
-package org.fundamentals.homework;
+package org.fundamentals.homework.animals;
+
+import org.fundamentals.homework.animals.behavior.MiceCatcher;
+import org.fundamentals.homework.animals.sounds.FrrSound;
 
 public class Hedgehog extends Animal implements MiceCatcher, FrrSound {
     public static final String NAME = "hedgehog";
