+++ fundamentals/src/org/fundamentals/homework/animals/sounds/AnimalSounds.java
@@ -1,4 +1,4 @@
-package org.fundamentals.homework;
+package org.fundamentals.homework.animals.sounds;
 
 public enum AnimalSounds {
     MEOW("meow"), BARK("woof"), FRR("frr");
