+++ fundamentals/src/org/fundamentals/homework/animals/sounds/FrrSound.java
@@ -1,4 +1,4 @@
-package org.fundamentals.homework;
+package org.fundamentals.homework.animals.sounds;
 
 public interface FrrSound extends SoundMaker {
     @Override
